+++ source/w_time.c
@@ -17,7 +17,7 @@
 static
 i32 lmvar frozen;
 
-cstr CanonTime(i32 type, u64 time) {
+cstr CanonTime(i32 type, i32 time) {
    noinit static
    char ft[64], st[64], dt[64], fmt[64];
